# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * SplitPay API — share-link metadata and indexer for on-chain bill splits
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ErrorResponse {
  message: string;
}

export type SplitType = (typeof SplitType)[keyof typeof SplitType];

export const SplitType = {
  equal: "equal",
  custom: "custom",
} as const;

export interface CreateSplitRequest {
  /** @maxLength 120 */
  title?: string;
  /** 0x address of the creator */
  creatorAddress: string;
  /** Numeric split id from the on-chain contract (stringified bigint) */
  onChainId: string;
  /** Total amount in 6-decimal USDC base units (microUSDC) as string */
  totalAmount: string;
  /**
   * @minimum 1
   * @maximum 100
   */
  participantCount: number;
  splitType: SplitType;
  /** Per-participant amounts (microUSDC) for custom splits */
  customAmounts?: string[];
  txHash: string;
}

export interface Split {
  id: string;
  title?: string;
  creatorAddress: string;
  onChainId: string;
  totalAmount: string;
  participantCount: number;
  splitType: SplitType;
  customAmounts?: string[];
  txHash: string;
  paidCount: number;
  createdAt: string;
}

export interface Payment {
  id: string;
  splitId: string;
  payerAddress: string;
  amount: string;
  txHash: string;
  paidAt: string;
}

export type SplitDetail = Split & {
  payments: Payment[];
};

export interface RecordPaymentRequest {
  payerAddress: string;
  /** microUSDC string */
  amount: string;
  txHash: string;
}

export interface Stats {
  totalSplits: number;
  /** microUSDC sum of all paid amounts */
  totalVolume: string;
  totalParticipantsPaid: number;
  activeSplits: number;
}

export interface AppConfig {
  chainId: number;
  chainName: string;
  rpcUrl: string;
  explorerUrl: string;
  usdcAddress: string;
  splitPayAddress: string;
}

export type ListRecentSplitsParams = {
  /**
   * @minimum 1
   * @maximum 50
   */
  limit?: number;
};