# lib/api-zod/src/generated/types/createSplitRequest.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * SplitPay API — share-link metadata and indexer for on-chain bill splits
 * OpenAPI spec version: 0.1.0
 */
import type { SplitType } from "./splitType";

export interface CreateSplitRequest {
  /** @maxLength 120 */
  title?: string;
  /** 0x address of the creator */
  creatorAddress: string;
  /** Numeric split id from the on-chain contract (stringified bigint) */
  onChainId: string;
  /** Total amount in 6-decimal USDC base units (microUSDC) as string */
  totalAmount: string;
  /**
   * @minimum 1
   * @maximum 100
   */
  participantCount: number;
  splitType: SplitType;
  /** Per-participant amounts (microUSDC) for custom splits */
  customAmounts?: string[];
  txHash: string;
}