# lib/api-zod/src/generated/types/recordPaymentRequest.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * SplitPay API — share-link metadata and indexer for on-chain bill splits
 * OpenAPI spec version: 0.1.0
 */

export interface RecordPaymentRequest {
  payerAddress: string;
  /** microUSDC string */
  amount: string;
  txHash: string;
}