# lib/api-zod/src/generated/types/split.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * SplitPay API — share-link metadata and indexer for on-chain bill splits
 * OpenAPI spec version: 0.1.0
 */
import type { SplitType } from "./splitType";

export interface Split {
  id: string;
  title?: string;
  creatorAddress: string;
  onChainId: string;
  totalAmount: string;
  participantCount: number;
  splitType: SplitType;
  customAmounts?: string[];
  txHash: string;
  paidCount: number;
  createdAt: Date;
}