# lib/api-zod/src/generated/types/stats.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * SplitPay API — share-link metadata and indexer for on-chain bill splits
 * OpenAPI spec version: 0.1.0
 */

export interface Stats {
  totalSplits: number;
  /** microUSDC sum of all paid amounts */
  totalVolume: string;
  totalParticipantsPaid: number;
  activeSplits: number;
}